Move BlockInformationPanel in dedicated dir (2)
